refactor(profil): extract nav bar context resolution into a concern
Shares the params[:context] resolution between both profile controllers.
The name avoids confusion with NavBarProfileConcern, which it overrides.
